+++ pog-task-manager/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "pog-task-manager" extension will be documented in th
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.1.1] - 2026-02-10
+
+### Fixed: Data Loss in `quickAdd`
+### Fixed: Config Overwrite & Unverified Download in `initPogTask`
+
 ## [1.1.0]
 
 - **YAML Transition**: Updated extension to support the new YAML-based project/module nested directory structure.